Router changes

+++ frontend/src/pages/user/Logout.jsx
@@ -4,9 +4,27 @@ import { useNavigate } from "react-router-dom";
 const Logout = () => {
   const navigate = useNavigate(); 
   const handleLogout = async () => {
-    await localStorage.clear();
-
-    navigate("/");
+    console.log("inside logout")
+    const response = await fetch(
+      "http://localhost:4000/api/auth/logout",
+      {
+        method: "POST",
+        credentials: "include",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }
+    );
+    const res=await response.json();
+    if(res.success){
+      localStorage.clear();
+      console.log(Cookies.get("accessToken"));
+      console.log(localStorage);
+      Cookies.remove("accessToken");
+      Cookies.remove("refreshToken");
+      navigate("/");
+    }
+   
   };
   useEffect(() => {
     handleLogout();
